requirements.txt
python-kasa
suntime~=1.2.5
timezonefinder~=6.2.0
tzdata
